Refactoring of login button to its own file. Minor improvements.
