+++ package.json
@@ -30,7 +30,7 @@
     "react-dom": "~18.2.0",
     "react-paginate": "~8.1.3",
     "react-perfect-scrollbar": "~1.5.8",
-    "react-router-dom": "~6.4.3",
+    "react-router-dom": "~6.8.1",
     "react-scripts": "~5.0.1",
     "sass": "~1.56.0",
     "react-twitter-embed": "~4.0.4",
